I implemented this wrong.  Apparently, time of year can be disjoint, as in Jan-Feb, May-Jul or stuff like that.  I'm going to reimplement it stored in a bitwise value.
In other words, Time of Year will be a binary number between 000000000000 and 111111111111.
Each digit will correspond to a month, where the first digit is January, the second February, and so on.
If the digit is set to 1, then the item can be interacted with in the corresponding month.
Otherwise, it can't be interacted with.

That may look confusing, but with some neat Javascript it should be fine.

It should be displayed as a series of checkboxes for each month, or something like that.