Added RewardTitle, RewardTitlePosition, and PassportTitle.
PassportTitles is the list of possible titles to be earned form completing Nook Miles milestones.

RewardTitles are the lists of PassportTitles for each milestone.

PassportTitles and RewardTitles are to be related in the following way:

PassportTitles have many RewardTitlePositions.
RewardTitlePositions belong to one PassportTitle.
RewardTitlePositions belong to one RewardTitle and act_as_list within its scope.
RewardTitles have many RewardTitlePositions.