Okay, BellValue is replaced when the modal opens now, but there is going to be a lot of work to do before a general CollectibleAttribute can be replaced.
I need to figure out a way to unintelligently traverse through all the CollectibleAttributes and get their values into the modal.

I am thinking a recursive function might work.  If the current CollectibleAttribute contains a div within it, it goes one level deeper.  If it is at the deepest level, it does something with the values.

Not sure yet how this will work.