Some explanation of CollectibleSort and CollectibleFilter:
Both have the same attributes:
Name
Keyword
Collectible_attribute_class_name

Here is how they work:
Each is a list of implemented sorts/filters.  If a sort/filter for the data in an IslandCollection exists, then we want to give the user the option to sort/filter by that sort/filter.  However, if it doesn't exist for a certain IslandCollection, we don't want it to show up there.

For instance, Fish have BellValues.
There is a CollectibleSort associated with BellValues.
Then when we render the page, we'll add a CollectibleSort to the select_options for the sorting dropdown.  We put its name in as the name of the option.  The keyword will be added to the url when that sort option is selected.  Actually, for sorts, the name and keyword will basically always be the same as the CollectibleAttribute's collectible_attribute_type.name
But, on the page for Villagers, none of them have a BellValue (for some reason?  Why can't I sell Villagers but I can sell other animals?)
Because no CollectibleAttributes with the class BellValue will appear on the page, the CollectibleSort for BellValue will not be found and will not be added to the sort options.

Basically filters work the same way.