Alright, some things I did and some things I need to do:
Did:
Added .default:boolean to IslandCollection.  This will be true if the IslandCollection should be given to each new User, and false if it is a custom collection created by a User.

I froze the attribute .default as well, so it will be created and then it cannot be changed.

Added the HABTM relationship between IslandCollections and Users.  Also made it so the index will show only the IslandCollections a User has access to.

To do:
Make sure that Users cannot send a request that sets .default as true for any IslandCollection.  That should only happen when seeding the database, then pretty much never again.

Change users#create to automatically assign all default IslandCollections.