I think I got the last todos complete.
Now if a User is signed in and tries to access users#new, they are redirected and told the user already exists.
However, the back button will bring the user back and allow them to create a new User, so I have to fix that next.

Also, the seed file has been updated to set IslandCollections is creates as default.

Finally, I got rid of the buttons to edit and delete default collections, because those are not needed.